+++ sdk/openai/azure-ai-openai/src/test/java/com/azure/ai/openai/OpenAIAsyncClientTest.java
@@ -25,6 +25,7 @@ import com.azure.ai.openai.models.ChatCompletionsToolSelectionPreset;
 import com.azure.ai.openai.models.ChatResponseMessage;
 import com.azure.ai.openai.models.ChatRole;
 import com.azure.ai.openai.models.Choice;
+import com.azure.ai.openai.models.CompleteUploadRequest;
 import com.azure.ai.openai.models.Completions;
 import com.azure.ai.openai.models.CompletionsFinishReason;
 import com.azure.ai.openai.models.CompletionsOptions;
@@ -39,6 +40,7 @@ import com.azure.ai.openai.models.OnYourDataApiKeyAuthenticationOptions;
 import com.azure.ai.openai.models.OnYourDataContextProperty;
 import com.azure.ai.openai.models.OpenAIFile;
 import com.azure.ai.openai.models.SpeechGenerationResponseFormat;
+import com.azure.ai.openai.models.UploadPart;
 import com.azure.core.exception.HttpResponseException;
 import com.azure.core.exception.ResourceNotFoundException;
 import com.azure.core.http.HttpClient;
@@ -58,11 +60,13 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static com.azure.ai.openai.TestUtils.DISPLAY_NAME_WITH_ARGUMENTS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -1725,5 +1729,62 @@ public class OpenAIAsyncClientTest extends OpenAIClientTestBase {
 //                })
                 .verifyComplete();
         }));
+    }
+
+    // Upload large file in parts
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testUploadLargesFilesInPartsOperations(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getOpenAIAsyncClient(httpClient, serviceVersion);
+        AtomicReference<String> uploadId = new AtomicReference<>();
+        uploadCreationRunner(createUploadRequest -> {
+            StepVerifier.create(client.createUpload(createUploadRequest))
+                    .assertNext(upload -> {
+                        assertNotNull(upload);
+                        assertNotNull(upload.getId());
+                        uploadId.set(upload.getId());
+                    }).verifyComplete();
+        });
+
+        addUploadPartRequestRunner((part1, part2) -> {
+            String uploadedId = uploadId.get();
+            assertNotNull(uploadedId);
+            StepVerifier.create(client.addUploadPart(uploadedId, part1)
+                .flatMap(uploadPartAdded -> {
+                    String uploadPartAddedId = uploadPartAdded.getId();
+                    assertNotNull(uploadPartAddedId);
+                    return client.addUploadPart(uploadedId, part2).zipWith(Mono.just(uploadPartAddedId));
+                })
+                .flatMap(tuple -> {
+                    UploadPart secondPart = tuple.getT1();
+                    assertNotNull(secondPart);
+                    String firstPartId = tuple.getT2();
+                    String secondPartId = secondPart.getId();
+                    assertNotEquals(firstPartId, secondPartId);
+                    CompleteUploadRequest completeUploadRequest = new CompleteUploadRequest(Arrays.asList(firstPartId, secondPartId));
+                    return client.completeUpload(uploadedId, completeUploadRequest);
+                }))
+                .assertNext(completeUpload -> {
+                    assertNotNull(completeUpload);
+                    assertEquals(uploadedId, completeUpload.getId());
+                }).verifyComplete();
+        });
+    }
+
+    @ParameterizedTest(name = DISPLAY_NAME_WITH_ARGUMENTS)
+    @MethodSource("com.azure.ai.openai.TestUtils#getTestParameters")
+    public void testCancelUploadLargesFilesInParts(HttpClient httpClient, OpenAIServiceVersion serviceVersion) {
+        client = getOpenAIAsyncClient(httpClient, serviceVersion);
+        uploadCreationRunner(createUploadRequest -> {
+            StepVerifier.create(client.createUpload(createUploadRequest)
+                    .flatMap(upload -> {
+                        assertNotNull(upload);
+                        assertNotNull(upload.getId());
+                        return client.cancelUpload(upload.getId());
+                    }))
+                .assertNext(cancelledId -> {
+                    assertNotNull(cancelledId);
+                }).verifyComplete();
+        });
     }
 }